+++ CHANGELOG.md
@@ -63,6 +63,7 @@ ssh_authorized_keys
 ### Removed
 
 - remove variable *host_ORACLE_ARCH*, *host_ORACLE_DATA* and *host_ORACLE_ROOT*
+- remove variable *host_env_config*
 
 ## [2.4.0] - 2023-03-29